Calculate private and public subnets for a VPC CIDR range
First grab as much space as possible for the private subnets (while leaving some space for the public subnets).
Then subtract the used subnets from the VPC CIDR range and return the combined range remaining.
Finally, using the remaining range grab as much space as possible for the public subnets.